requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
Werkzeug==2.3.7
requests==2.31.0
PyMySQL==1.1.0
cryptography==41.0.4
python-socketio==5.8.0
python-engineio==4.7.1
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3
python-dateutil==2.8.2